+++ SPMeta2/SPMeta2.Regression.Tests/Impl/Scenarios/AppScenariosTest.cs
@@ -26,7 +26,7 @@ using SPMeta2.Regression.Utils;
 namespace SPMeta2.Regression.Tests.Impl.Scenarios
 {
     [TestClass]
-    public class AppScenariousTest : SPMeta2RegresionScenarioTestBase
+    public class AppScenariosTest : SPMeta2RegresionScenarioTestBase
     {
         #region internal
