+++ SPMeta2/SPMeta2.Regression.Tests/Impl/Scenarios/SandboxSolutionScenariosTest.cs
@@ -1,4 +1,9 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using System.Diagnostics;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SPMeta2.Containers;
+using SPMeta2.Containers.Services;
+using SPMeta2.Definitions;
+using SPMeta2.Exceptions;
 using SPMeta2.Regression.Tests.Base;
 using SPMeta2.Regression.Tests.Impl.Scenarios.Base;
 using System;
@@ -6,11 +11,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
+using SPMeta2.Syntax.Default;
+using System.IO;
 
 namespace SPMeta2.Regression.Tests.Impl.Scenarios
 {
     [TestClass]
-    public class ResetRoleInheritanceScenariousTest : SPMeta2RegresionScenarioTestBase
+    public class SandboxSolutionScenariosTest : SPMeta2RegresionScenarioTestBase
     {
         #region internal
 
@@ -30,13 +37,7 @@ namespace SPMeta2.Regression.Tests.Impl.Scenarios
 
         #region default
 
-        // web scope, TODO
-
-        // list scope, TODO
-
-        // folder scope, TODO
-
-        // list item scope, TODO
+       
 
         #endregion
     }
