+++ SPMeta2/SPMeta2.Regression.Tests/Impl/Scenarios/UniqueContentTypeFieldsScenariosTest.cs
@@ -14,7 +14,7 @@ using SPMeta2.Syntax.Default;
 namespace SPMeta2.Regression.Tests.Impl.Scenarios
 {
     [TestClass]
-    public class UniqueContentTypeFieldsScenariousTest : SPMeta2RegresionScenarioTestBase
+    public class UniqueContentTypeFieldsScenariosTest : SPMeta2RegresionScenarioTestBase
     {
         #region internal
