+++ SPMeta2/SPMeta2.Regression.Tests/SPMeta2.Regression.Tests.csproj
@@ -180,11 +180,11 @@
     <Compile Include="Impl\Random\RandomDefinitionTest.cs" />
     <Compile Include="Impl\RegressionAPI\RegressionAPITests.cs" />
     <Compile Include="Impl\Reports\DefinitionReportTests.cs" />
-    <Compile Include="Impl\Scenarios\ComposedLookItemLinkScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\ComposedLookItemLinkScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\DocumentSetScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\DeleteQuickLaunchNavigationNodesScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\DeleteTopNavigationNodesScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\Fields\OutcomeChoiceFieldScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\Fields\OutcomeChoiceFieldScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\CorePropertyScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\FarmSolutionScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\DesignPackageScenariosTest.cs" />
@@ -193,53 +193,53 @@
     <Compile Include="Impl\Scenarios\WorkflowAssociationScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\MetadataNavigationSettingsScenarios.cs" />
     <Compile Include="Impl\Scenarios\ManagedPropertiesScenarios.cs" />
-    <Compile Include="Impl\Scenarios\SupportedUICultureScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\AppScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\SupportedUICultureScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\AppScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\Base\SPMeta2RegresionScenarioTestBase.cs" />
-    <Compile Include="Impl\Scenarios\Fields\DependentLookupFieldScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\Fields\CalculatedFieldScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\Fields\UserFieldScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\Fields\DependentLookupFieldScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\Fields\CalculatedFieldScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\Fields\UserFieldScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\ListViewWebPartScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\OOTBPageScenarios.cs" />
-    <Compile Include="Impl\Scenarios\ResetRoleInheritanceScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\ContentTypeFieldLinkScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\ContentTypeLinkScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\ResetRoleInheritanceScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\ContentTypeFieldLinkScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\ContentTypeLinkScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\ContentTypeScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\EventReceiverScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\Fields\ChoiceFieldScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\Fields\LookupFieldScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\Fields\ChoiceFieldScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\Fields\LookupFieldScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\JobScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\AuditSettingsScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\ListItemValueScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\AuditSettingsScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\ListItemValueScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\MasterPageSettingsDefinitionScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\PageLayoutAndSiteTemplateSettingsScenarios.cs" />
     <Compile Include="Impl\Scenarios\PublishingPageLayoutScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\PublishingPageScenarios.cs" />
     <Compile Include="Impl\Scenarios\ReusableItemScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\SiteScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\SandboxSolutionScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\SiteScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\SandboxSolutionScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\QuickLaunchNavigationNodeScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\SecurityGroupScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\RootWebScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\TaxonomyFieldScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\TaxonomyTermStoreScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\TaxonomyTermScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\TaxonomyTermSetScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\TaxonomyTermGroupScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\TaxonomyFieldScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\TaxonomyTermStoreScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\TaxonomyTermScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\TaxonomyTermSetScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\TaxonomyTermGroupScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\TopNavigationNodeScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\UserCustomActionScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\UserCustomActionScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\FieldScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\FolderScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\ListFieldLinkScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\ListItemScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\FolderScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\ListFieldLinkScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\ListItemScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\ListScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\FeatureScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\FeatureScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\PropertyBagScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\TaxonomyScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\ListViewScenariousTest.cs" />
-    <Compile Include="Impl\Scenarios\ModuleFileScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\TaxonomyScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\ListViewScenariosTest.cs" />
+    <Compile Include="Impl\Scenarios\ModuleFileScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\SP2013WorkflowScenariosTest.cs" />
-    <Compile Include="Impl\Scenarios\UniqueContentTypeFieldsScenariousTest.cs" />
+    <Compile Include="Impl\Scenarios\UniqueContentTypeFieldsScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\WebPartPageScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\Webparts\ContentEditorWebPartScenariosTest.cs" />
     <Compile Include="Impl\Scenarios\Webparts\XsltListViewWebPartScenariosTest.cs" />
